# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Silvae.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5089/"
    }
  ],
  "paths": {
...
    "/api/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "GetDocuments",
        "parameters": [
          {
            "name": "worksiteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "operationId": "UploadDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file",
                  "title",
                  "category"
                ],
                "type": "object",
                "allOf": [
                  {
                    "type": "object",
                    "properties": {
                      "file": {
                        "$ref": "#/components/schemas/IFormFile"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "category": {
                        "type": "string"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "worksiteId": {
                        "type": "string",
                        "format": "uuid"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "issuedOn": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "expiresOn": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/documents/{documentId}/content": {
      "get": {
        "tags": [
          "Documents"
        ],
        "operationId": "DownloadDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/documents/{documentId}": {
      "delete": {
        "tags": [
          "Documents"
        ],
        "operationId": "DeleteDocument",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentDto": {
        "required": [
          "id",
          "worksiteId",
          "title",
          "category",
          "issuedOn",
          "expiresOn",
          "fileName",
          "contentType",
          "sizeBytes",
          "isExpired",
          "uploadedBy",
          "uploadedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "worksiteId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "issuedOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "expiresOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isExpired": {
            "type": "boolean"
          },
          "uploadedBy": {
            "type": "string",
            "format": "uuid"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
    }
  },
...
}